Refactor customer related parser test
