+++ src/test/java/seedu/address/logic/parser/customer/CustomerDeleteCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.customer;
 
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
@@ -8,6 +8,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.customer.CustomerDeleteCommand;
+import seedu.address.logic.parser.CustomerDeleteCommandParser;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path
